requirements.txt
Django==6.0
Pillow==12.0.0
qrcode==8.2
unidecode==1.3.8

# Celery (background tasks)
celery==5.3.6
sqlalchemy==2.0.23

# Google Business Profile API
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# Environment variables
python-dotenv==1.0.0
